Move Twitter related scripts out of candidates app
